+++ impl-base/src/main/java/org/jboss/shrinkwrap/impl/base/io/tar/TarBzInputStream.java
@@ -22,11 +22,12 @@ import java.io.InputStream;
 import org.jboss.shrinkwrap.impl.base.io.tar.bzip.BZip2CompressorInputStream;
 
 /**
- * The {@link TarBzInputStream} reads a UNIX TAR archive, further encoded in GZIP compresssion, as an InputStream.
+ * The {@link TarBzInputStream} reads a UNIX TAR archive, further encoded in BZIP compresssion, as an InputStream.
  * Methods are provided to position at each successive entry in the archive, and the read each entry as a normal input
  * stream using read().
  *
  * @author <a href="mailto:andrew.rubinger@jboss.org">ALR</a>
+ * @author <a href="mailto:ts@bee.kz">Tair Sabirgaliev</a>
  */
 public class TarBzInputStream extends TarInputStream {
