+++ impl-base/src/main/java/org/jboss/shrinkwrap/impl/base/io/tar/bzip/BZip2CompressorInputStream.java
@@ -30,10 +30,59 @@ import java.io.InputStream;
 /**
  * An input stream that decompresses from the BZip2 format to be read as any other stream.
  *
+ * <p><em>Forked from <a href="http://commons.apache.org/proper/commons-compress/">
+ * Apache Commons-Compress</a> with possible changes</em></p>
+ *
  * @NotThreadSafe
  */
-public class BZip2CompressorInputStream extends CompressorInputStream implements
-                                                                          BZip2Constants {
+public class BZip2CompressorInputStream extends InputStream implements BZip2Constants {
+
+    private long bytesRead = 0;
+
+    /**
+     * Increments the counter of already read bytes.
+     * Doesn't increment if the EOF has been hit (read == -1)
+     *
+     * @param read the number of bytes read
+     *
+     * @since 1.1
+     */
+    protected void count(int read) {
+        count((long) read);
+    }
+
+    /**
+     * Increments the counter of already read bytes.
+     * Doesn't increment if the EOF has been hit (read == -1)
+     *
+     * @param read the number of bytes read
+     */
+    protected void count(long read) {
+        if(read != -1) {
+            bytesRead = bytesRead + read;
+        }
+    }
+
+    /**
+     * Returns the current number of bytes read from this stream.
+     * @return the number of read bytes
+     * @deprecated this method may yield wrong results for large
+     * archives, use #getBytesRead instead
+     */
+    @Deprecated
+    public int getCount() {
+        return (int) bytesRead;
+    }
+
+    /**
+     * Returns the current number of bytes read from this stream.
+     * @return the number of read bytes
+     *
+     * @since 1.1
+     */
+    public long getBytesRead() {
+        return bytesRead;
+    }
 
     /**
      * Index of the last char in the block, so the block size == last + 1.
